requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
supabase==1.2.0
Flask-CORS==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0